* Fixed exit portals * Added auto conversion of materials using PNGs with alpha channel to null blends. * Fixed some warnings * Changed include in OpenCL.h back to #include <CL/cl.h>
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@2002 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/Parser.cpp
@@ -477,7 +477,7 @@ void Parser::doUnitTests()
 		double y = 0;
 		for(int i=0; i<1000000; ++i)
 		{
-			Parser p(text.c_str(), text.size());
+			Parser p(text.c_str(), (unsigned int)text.size());
 
 			double x;
 			p.parseDouble(x);
@@ -498,7 +498,7 @@ void Parser::doUnitTests()
 		double y = 0;
 		for(int i=0; i<1000000; ++i)
 		{
-			p.reset(text.c_str(), text.size());
+			p.reset(text.c_str(), (unsigned int)text.size());
 
 			double x;
 			p.parseDouble(x);
